moving utxoPersistenceService interface, UtxoTransactionReader interface
