+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/utxo/impl/request/handlers/UtxoFindTransactionsWithStatusCreatedBetweenTimeRequestHandler.kt
@@ -7,7 +7,7 @@ import net.corda.data.persistence.EntityResponse
 import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
 import net.corda.ledger.common.data.transaction.TransactionStatus.Companion.toTransactionStatus
 import net.corda.ledger.persistence.common.RequestHandler
-import net.corda.ledger.persistence.utxo.UtxoPersistenceService
+import net.corda.ledger.libs.utxo.UtxoPersistenceService
 import net.corda.messaging.api.records.Record
 import net.corda.v5.application.serialization.SerializationService
 import java.nio.ByteBuffer
