+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/utxo/impl/request/handlers/UtxoFindUnconsumedStatesByTypeRequestHandler.kt
@@ -4,7 +4,7 @@ import net.corda.data.flow.event.external.ExternalEventContext
 import net.corda.data.ledger.persistence.FindUnconsumedStatesByType
 import net.corda.ledger.persistence.common.RequestHandler
 import net.corda.ledger.persistence.utxo.UtxoOutputRecordFactory
-import net.corda.ledger.persistence.utxo.UtxoPersistenceService
+import net.corda.ledger.libs.utxo.UtxoPersistenceService
 import net.corda.messaging.api.records.Record
 import net.corda.sandboxgroupcontext.SandboxGroupContext
 import net.corda.v5.ledger.utxo.ContractState
