+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/utxo/impl/request/handlers/UtxoIncrementTransactionRepairAttemptCountRequestHandler.kt
@@ -6,7 +6,7 @@ import net.corda.data.ledger.persistence.IncrementTransactionRepairAttemptCount
 import net.corda.data.persistence.EntityResponse
 import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
 import net.corda.ledger.persistence.common.RequestHandler
-import net.corda.ledger.persistence.utxo.UtxoPersistenceService
+import net.corda.ledger.libs.utxo.UtxoPersistenceService
 import net.corda.messaging.api.records.Record
 
 class UtxoIncrementTransactionRepairAttemptCountRequestHandler(
