+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/utxo/impl/request/handlers/UtxoPersistTransactionRequestHandler.kt
@@ -3,9 +3,9 @@ package net.corda.ledger.persistence.utxo.impl.request.handlers
 import net.corda.data.flow.event.external.ExternalEventContext
 import net.corda.ledger.persistence.common.RequestHandler
 import net.corda.ledger.persistence.utxo.UtxoOutputRecordFactory
-import net.corda.ledger.persistence.utxo.UtxoPersistenceService
+import net.corda.ledger.libs.utxo.UtxoPersistenceService
 import net.corda.ledger.persistence.utxo.UtxoTokenObserverMap
-import net.corda.ledger.persistence.utxo.UtxoTransactionReader
+import net.corda.ledger.libs.utxo.UtxoTransactionReader
 import net.corda.ledger.persistence.utxo.impl.TokenStateObserverContextImpl
 import net.corda.messaging.api.records.Record
 import net.corda.v5.application.crypto.DigestService
